Enabled font/text colouring Fixed GU Flag cache bug (remember to never use sceGuEnable/disable directly.....)

+++ include/graphics/UIRender.h
@@ -29,7 +29,7 @@ namespace UI {
     void drawSprite(int x, int y, int u, int v, uint32_t color, float xScale, float yScale);
 
     void drawString(int x, int y, uint32_t color, float scale, float yScale, const std::string& text);
-    void drawRect(float x, float y, float w, float h, unsigned int colour);
+    void drawRect(short x, short y, short w, short h, unsigned int colour);
     void drawTri(float x1, float y1, float x2, float y2, float x3, float y3, unsigned int color);
 
     void drawHealthBar(float x, float y, float w, float h, int health, int maxHealth);
